moves roi_masks up the tree, points decrosstalk to it
