+++ src/ophys_etl/utils/roi_masks.py
